backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
SQLAlchemy==2.0.36
pydantic-settings==2.6.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
reportlab==4.2.5
psycopg[binary]==3.2.3

# testing
pytest==8.3.2
httpx==0.27.0
pytest-asyncio==0.23.8

